requirements.txt
Flask
Flask-Login
Werkzeug
psycopg2-binary
python-dotenv